Fix typo in workflow
You can't run a file that doesn't exist! The rule with the name of
"Build x86_64" was running this:

    {insignificant} && python3 smeltile.py

I looked trough the repo and couldn't find any "smeltile.py" file, but
the package "smelt3" has a configuration that requires a "smeltfile.py"
file, I also found a smeltfile.py file in the filesystem, so I went
ahead with that.