Debug config
I believe there was a substantial issue in the "config/config.sh"
directory. Since all of the .asm files the config links the profile
files with are by default abiding to the "default" profiles rules.
Therefore I added a check to see if the profile is default. If so,
we exit.